src/requirements.txt
PyPDF2
pdf2image
Pillow
pytesseract
docx
pdf2docx
